requirements.txt
numpy
gradescope-utils
matplotlib
torch
tensorboardX
tensorboard
torchvision
pyyaml
tqdm
